backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-core
langchain-groq
langchain-tavily
langgraph
pydantic
python-multipart